Squashed commit of the following:
    MyTablesLambdaALBListenerRule ==> MyTablesLambdaTG ==> {{baseUrl}}/api/tables

	A powershell script to display for all of non-default resources in the current account using aws cli: ./list-all-non-default-resources.ps1

+++ scripts/dev-build-deploy-delete.ps1
@@ -1,7 +1,14 @@
+param (
+   [bool]$skipBuild = $false,
+   [bool]$deleteTask = $false
+)
+
 . .\Get-ElapsedTimeFormatted.ps1
 $startTime = Get-Date
 
-./dev-build-deploy.ps1
+if (-not $skipBuild) {
+   ./dev-build-deploy.ps1
+}
 
 $commonConstants = ./common-constants.ps1
 
@@ -16,18 +23,50 @@ Write-Output "`n$(Get-Date -Format 'HH:mm:ss'), elapsed $formattedElapsedTime : 
 Write-Output "`n$(Get-Date -Format 'HH:mm:ss'), elapsed $formattedElapsedTime : Testing all HTTP requests deployed by the stack ..."
 ./test-web-api.ps1
 
-$formattedElapsedTime = Get-ElapsedTimeFormatted -startTime $startTime
-Write-Output "`n$(Get-Date -Format 'HH:mm:ss'), elapsed $formattedElapsedTime : Deleting stack '$($commonConstants.stackName)' ..."
-aws cloudformation delete-stack --stack-name $commonConstants.stackName
-aws cloudformation wait stack-delete-complete --stack-name $commonConstants.stackName
+# $scriptsFolder = (Get-Location).Path
+# Set-Location ../
+# ./backend/sql-tables/test-pg-local.ps1
+# Set-Location ${scriptsFolder}
+
+if (-not $deleteTask) {
+   $validInput = $false
+   while (-not $validInput) {
+       $message = "Stack deployed. Continue to [d]elete the Stack or E[x]it ?"
+       Write-Host $message  -ForegroundColor White -BackgroundColor DarkGray
+       $userInput = Read-Host ">"
+
+       if ($userInput -eq 'x' -or $userInput -eq 'X') {
+           Write-Host "Exiting the script."
+           exit
+       } elseif ($userInput -eq 'd' -or $userInput -eq 'D') {
+           Write-Host "Continuing to delete the stack..."
+           $deleteTask = $true
+           $validInput = $true
+       } else {
+           Write-Host "Invalid input. Please enter 'd' or 'x'."
+       }
+   }
+}
+
+if ($deleteTask) {
+   $formattedElapsedTime = Get-ElapsedTimeFormatted -startTime $startTime
+   Write-Output "`n$(Get-Date -Format 'HH:mm:ss'), elapsed $formattedElapsedTime : Deleting stack '$($commonConstants.stackName)' ..."
+   aws cloudformation delete-stack --stack-name $commonConstants.stackName
+   aws cloudformation wait stack-delete-complete --stack-name $commonConstants.stackName
+
+   $formattedElapsedTime = Get-ElapsedTimeFormatted -startTime $startTime
+   Write-Output "`n$(Get-Date -Format 'HH:mm:ss'), elapsed $formattedElapsedTime : Listing ECS clusters, services, tasks and containers ..."
+   ./List-ECS.ps1 -region $commonConstants.region
+
+   $formattedElapsedTime = Get-ElapsedTimeFormatted -startTime $startTime
+   Write-Output "`n$(Get-Date -Format 'HH:mm:ss'), elapsed $formattedElapsedTime : Listing all subnets ..."
+   ./List-Subnets.ps1 -region $commonConstants.region
+}
 
 $formattedElapsedTime = Get-ElapsedTimeFormatted -startTime $startTime
-Write-Output "`n$(Get-Date -Format 'HH:mm:ss'), elapsed $formattedElapsedTime : Listing ECS clusters, services, tasks and containers ..."
-./List-ECS.ps1 -region $commonConstants.region
-
-$formattedElapsedTime = Get-ElapsedTimeFormatted -startTime $startTime
-Write-Output "`n$(Get-Date -Format 'HH:mm:ss'), elapsed $formattedElapsedTime : Listing all subnets ..."
-./List-Subnets.ps1 -region $commonConstants.region
+Write-Output "`n$(Get-Date -Format 'HH:mm:ss'), elapsed $formattedElapsedTime : Listing all non-default resources ..."
+./list-all-non-default-resources.ps1
 
 $formattedElapsedTime = Get-ElapsedTimeFormatted -startTime $startTime
 Write-Output "`n$(Get-Date -Format 'HH:mm:ss'), elapsed $formattedElapsedTime : Completed."
+
